+++ lib/wardrobe/middleware.rb
@@ -1,0 +1,15 @@
+module Wardrobe
+  class Middleware
+    attr_reader :name, :block, :use_if
+
+    def initialize(name, block, use_if)
+      @name = name
+      @block = block
+      @use_if = use_if
+    end
+
+    def <=>(other)
+      priority <=> other.priority
+    end
+  end
+end