+++ lib/wardrobe/root_config.rb
@@ -8,7 +8,7 @@ module Wardrobe
     plugin :default
     attribute :coerce, Boolean, default: true
     attribute :logger, Object, default: ->() { Logger.new(STDOUT) }
-    attribute :default_plugins, Set[Symbol], default: Set.new([:coercible])
+    attribute :default_plugins, Set[Symbol], default: ->() { Set.new([:coercible]) }
 
     def register_default_plugin(name)
       raise 'error' unless Wardrobe.plugins.key?(name)